SPARK-44646 Reduce usage of log4j2 core classes
