+++ resource-managers/yarn/src/main/scala/org/apache/spark/deploy/yarn/SparkRackResolver.scala
@@ -27,7 +27,6 @@ import org.apache.hadoop.net._
 import org.apache.hadoop.util.ReflectionUtils
 import org.apache.hadoop.yarn.util.RackResolver
 import org.apache.logging.log4j.{Level, LogManager}
-import org.apache.logging.log4j.core.Logger
 
 import org.apache.spark.internal.Logging
 
@@ -40,8 +39,8 @@ private[spark] class SparkRackResolver(conf: Configuration) extends Logging {
 
   // RackResolver logs an INFO message whenever it resolves a rack, which is way too often.
   val logger = LogManager.getLogger(classOf[RackResolver])
-  if (logger.getLevel != Level.WARN) {
-    logger.asInstanceOf[Logger].setLevel(Level.WARN)
+  if (logger.getLevel != Level.WARN && logger.getLevel != Level.ERROR) {
+    Logging.setLogLevel(Level.WARN)
   }
 
   private val dnsToSwitchMapping: DNSToSwitchMapping = {
